Fix: Don't read 'plane#_yolk_' dataRefs if TCAS override, fixes #44
