+++ XPMP2-Sample/lib/XPMP2.framework/Versions/Current
@@ -1,1 +1,1 @@
-2.3
+2.4
